+++ apworld/brotato/options.py
@@ -10,38 +10,38 @@ from .constants import (
     MAX_NORMAL_CRATE_DROP_GROUPS,
     MAX_NORMAL_CRATE_DROPS,
     MAX_RARE_UPGRADES,
-    MAX_SHOP_LOCATIONS_PER_TIER,
     MAX_SHOP_SLOTS,
     MAX_UNCOMMON_UPGRADES,
     NUM_CHARACTERS,
     NUM_WAVES,
-    ItemRarity,
 )
 
 
 class NumberRequiredWins(Range):
-    """The number of characters you must complete runs with to win."""
+    """The number of runs you need to win to complete your goal.
+
+    Each win must be done with a different character.
+    """
 
     range_start = 1
     range_end = NUM_CHARACTERS
 
-    display_name = "Number of runs required"
     default = 10
+    display_name = "Wins Required"
 
 
 class StartingCharacters(TextChoice):
     """Determines your set of starting characters.
 
-    Default: Start with Well Rounded, Brawler, Crazy, Ranger and Mage.
-
-    Shuffle: Start with a random selection of characters.
+    * Default: Start with Well Rounded, Brawler, Crazy, Ranger and Mage.
+    * Shuffle: Start with a random selection of characters.
     """
 
     option_default_characters = 0
     option_random_characters = 1
 
-    display_name = "Starting characters"
     default = 0
+    display_name = "Starting Characters"
 
 
 class NumberStartingCharacters(Range):
@@ -50,12 +50,15 @@ class NumberStartingCharacters(Range):
     range_start = 1
     range_end = NUM_CHARACTERS
 
-    display_name = "Number of starting characters"
     default = 5
+    display_name = "Number of Starting Characters"
 
 
 class WavesPerCheck(Range):
-    """How many waves to win to receive a check. Smaller values mean more frequent checks."""
+    """How many waves to win to receive a check.
+
+    1 means every wave is a check, 2 means every other wave, etc.
+    """
 
     # We'd make the start 1, but the number of items sent when the game is released is
     # so large that the resulting ReceivedItems command is bigger than Godot 3.5's
@@ -63,14 +66,12 @@ class WavesPerCheck(Range):
     range_start = 1
     range_end = NUM_WAVES
 
-    display_name = "Waves per check"
     default = 10
+    display_name = "Waves Per Check"
 
 
 class NumberCommonCrateDropLocations(Range):
-    """The number of loot crate locations.
-
-    This replaces the loot crate drops in-game with an Archipelago item which must be picked up.
+    """Replaces the in-game loot crate drops with an Archipelago item which must be picked up to generate a check.
 
     How the drops are made available and how many are needed to make a check are controlled by the next two settings.
     """
@@ -78,30 +79,31 @@ class NumberCommonCrateDropLocations(Range):
     range_start = 0
     range_end = MAX_NORMAL_CRATE_DROPS
 
-    display_name = "Number of normal crate drop locations"
     default = 25
+    display_name = "Loot Crate Locations"
 
 
 class NumberCommonCrateDropsPerCheck(Range):
-    """The number of loot crates needed to check a location.
+    """The number of common loot crates which need to be picked up to count as a check.
 
-    1 means every loot crate pickup gives a check,
-    2 means every other loot crate,
-    etc.
+    1 means every crate is a check, 2 means every other crate, etc.
     """
 
-    range_start = 0
+    range_start = 1
     range_end: int = MAX_NORMAL_CRATE_DROPS
 
-    display_name: str = "Loot Crates per Check"
+    default = 2
+    display_name: str = "Crate Pickup Step"
 
 
 class NumberCommonCrateDropGroups(Range):
     """The number of groups to separate loot crate locations into.
 
-    Once you check all the locations in a group, the randomizer will not drop more loot crate Archipelago items until you win more runs.
+    Once you check all the locations in a group, the randomizer will not drop more loot crate Archipelago items until
+    you win more runs.
 
-    The number of loot crate locations will be evenly split among the groups, and the groups will be evenly spread out over the number of wins you choose.
+    The number of loot crate locations will be evenly split among the groups, and the groups will be evenly spread out
+    over the number of wins you choose.
 
     Set to 1 to make all loot crate locations available from the start.
     """
@@ -109,14 +111,13 @@ class NumberCommonCrateDropGroups(Range):
     range_start = 1
     range_end: int = MAX_NORMAL_CRATE_DROP_GROUPS
 
-    display_name: str = "Loot Crate Groups"
     default = 1
+    display_name: str = "Loot Crate Groups"
 
 
 class NumberLegendaryCrateDropLocations(Range):
-    """The number of legendary loot crate locations.
-
-    This replaces the legendary loot crate drops in-game with an Archipelago item which must be picked up.
+    """Replaces the in-game legendary loot crate drops with an Archipelago item which must be picked up to generate a
+    check.
 
     How the drops are made available and how many are needed to make a check are controlled by the next two settings.
     """
@@ -124,37 +125,40 @@ class NumberLegendaryCrateDropLocations(Range):
     range_start = 0
     range_end: int = MAX_LEGENDARY_CRATE_DROPS
 
-    display_name: str = "Number of legendary crate drop locations"
     default = 5
+    display_name: str = "Legendary Loot Crate Locations"
 
 
 class NumberLegendaryCrateDropsPerCheck(Range):
-    """The number of legendary loot crates needed to check a location.
+    """The number of legendary loot crates which need to be picked up to count as a check.
 
-    1 means every loot crate pickup gives a check, 2 means every other loot crate, etc.
+    1 means every crate is a check, 2 means every other crate, etc.
     """
 
-    range_start = 0
+    range_start = 1
     range_end: int = MAX_NORMAL_CRATE_DROPS
-    default = 1
 
-    display_name: str = "Loot crates per check"
+    default = 1
+    display_name: str = "Legendary Loot Crate Pickup Step"
 
 
 class NumberLegendaryCrateDropGroups(Range):
     """The number of groups to separate legendary loot crate locations into.
 
-    Once you check all the locations in a group, the randomizer will not drop more legendary loot crate Archipelago items until you win more runs.
+    Once you check all the locations in a group, the randomizer will not drop more legendary loot crate Archipelago
+    items until you win more runs.
 
-    The number of loot crate locations will be evenly split among the groups, and the groups will be evenly spread out over the number of wins you choose.
+    The number of loot crate locations will be evenly split among the groups, and the groups will be evenly spread out
+    over the number of wins you choose.
 
     Set to 1 to make all legendary loot crate locations available from the start.
     """
 
     range_start = 1
     range_end: int = MAX_LEGENDARY_CRATE_DROP_GROUPS
+
     default = 1
-    display_name: str = "Loot Crate Groups"
+    display_name: str = "Legendary Loot Crate Groups"
 
 
 class ItemWeights(Choice):
@@ -163,11 +167,12 @@ class ItemWeights(Choice):
     For every common crate drop location, a Brotato weapon/item will be added to the pool. This controls how the item
     tiers are chosen.
 
-    Note that legendary crate drop locations will ALWAYS add a legendary item to the pool, which is in addition to any legendary items added by common crate locations.
+    Note that legendary crate drop locations will ALWAYS add a legendary item to the pool, which is in addition to any
+    legendary items added by common crate locations.
 
-    - Default: Use the game's normal distribution. Equivalent to setting the custom weights to 100/60/25/8.
-    - Chaos: Each tier has a has a random weight.
-    - Custom: Use the custom weight options below.
+    * Default: Use the game's normal distribution. Equivalent to setting the custom weights to 100/60/25/8.
+    * Chaos: Each tier has a has a random weight.
+    * Custom: Use the custom weight options below.
     """
 
     option_default = 0
@@ -183,8 +188,8 @@ class CommonItemWeight(Range):
     range_start = 0
     range_end = 100
 
-    display_name = "Common Items"
     default = 100
+    display_name = "Common Items"
 
 
 class UncommonItemWeight(Range):
@@ -193,8 +198,8 @@ class UncommonItemWeight(Range):
     range_start = 0
     range_end = 100
 
-    display_name = "Uncommon Items"
     default = 60
+    display_name = "Uncommon Items"
 
 
 class RareItemWeight(Range):
@@ -203,61 +208,62 @@ class RareItemWeight(Range):
     range_start = 0
     range_end = 100
 
-    display_name = "Rare Items"
     default = 25
+    display_name = "Rare Items"
 
 
 class LegendaryItemWeight(Range):
     """The weight of Legendary/Tier 4/Red items in the pool.
 
-    Note that this is for common crate drop locations only. An additional number of legendary items is also added for each legendary crate drop location.
+    Note that this is for common crate drop locations only. An additional number of legendary items is also added for
+    each legendary crate drop location.
     """
 
     range_start = 0
     range_end = 100
 
-    display_name = "Legendary Items"
     default = 8
+    display_name = "Legendary Items"
 
 
 class NumberCommonUpgrades(Range):
-    """The normal of level 1 upgrades to include in the item pool."""
+    """The number of Common/Tier 1/White upgrades to include in the item pool."""
 
     range_start = 0
     range_end: int = MAX_COMMON_UPGRADES
 
-    display_name: str = "Number of level 1 upgrades"
     default = 15
+    display_name: str = "Common Upgrades"
 
 
 class NumberUncommonUpgrades(Range):
-    """The normal of level 2 upgrades to include in the item pool."""
+    """The number of Uncommon/Tier 2/Blue upgrades to include in the item pool."""
 
     range_start = 0
     range_end: int = MAX_UNCOMMON_UPGRADES
 
-    display_name: str = "Number of level 2 upgrades"
     default = 10
+    display_name: str = "Uncommon Upgrades"
 
 
 class NumberRareUpgrades(Range):
-    """The normal of level 3 upgrades to include in the item pool."""
+    """The number of Rare/Tier 3/Purple upgrades to include in the item pool."""
 
     range_start = 0
     range_end: int = MAX_RARE_UPGRADES
 
-    display_name: str = "Number of level 3 upgrades"
     default = 5
+    display_name: str = "Rare Upgrades"
 
 
 class NumberLegendaryUpgrades(Range):
-    """The normal of level 4 upgrades to include in the item pool."""
+    """The number of Legendary/Tier 4/Red upgrades to include in the item pool."""
 
     range_start = 0
     range_end = MAX_LEGENDARY_UPGRADES
 
-    display_name = "Number of level 4 upgrades"
     default = 5
+    display_name = "Legendary Upgrades"
 
 
 class StartingShopSlots(Range):
@@ -265,17 +271,9 @@ class StartingShopSlots(Range):
 
     range_start = 0
     range_end: int = MAX_SHOP_SLOTS
-    display_name: str = "Starting shop slots"
+
     default = 4
-
-
-class NumberShopItems(Range):
-    """The number of items to place in the shop"""
-
-    range_start = 0
-    range_end: int = MAX_SHOP_LOCATIONS_PER_TIER[ItemRarity.COMMON]
-    display_name: str = "Shop items"
-    default = 10
+    display_name: str = "Starting Shop Slots"
 
 
 @dataclass
@@ -300,4 +298,3 @@ class BrotatoOptions(PerGameCommonOptions):
     num_rare_upgrades: NumberRareUpgrades
     num_legendary_upgrades: NumberLegendaryUpgrades
     num_starting_shop_slots: StartingShopSlots
-    num_shop_items: NumberShopItems